Outsourced core of Location to Url class
