+++ AngleSharp/AngleSharp.csproj
@@ -296,6 +296,7 @@
     <Compile Include="Foundation\LocationExtensions.cs" />
     <Compile Include="Foundation\Setup\CopyConfiguration.cs" />
     <Compile Include="Foundation\Structures\Priority.cs" />
+    <Compile Include="Foundation\Structures\Url.cs" />
     <Compile Include="Foundation\StyleExtensions.cs" />
     <Compile Include="Foundation\TextSource.cs" />
     <Compile Include="Infrastructure\CssStyleEngine.cs" />
